Use UserNotifications to create local notifications
a15e038a adds support for enumerating/dismissing delivered notifications
by ID, but UserNotifications notifications and UILocalNotification
notifications have a different set of IDs (a notification created using
UILocalNotification will have an ID that does not match the
notification's ID in the UserNotifications world and hence cannot be
used to dismiss that notification). This commit creates the
notifications using UserNotifications if the user is on iOS 10.