Fix EventsRegistry.registerNotificationOpened typing (#499)
`EventsRegistry.registerNotificationOpened` registers a `callback` that expects a `response: NotificationResponse` object as the first parameter. In the registered callback, we can expect `response.notificaiton` to be the notification opened.
However, `NativeEventsReceiver.registerNotificationOpened` wants a `callback` that expects a `notification: Notification` object instead. So we are passing back `notification` as the `response` which will cause the callback code to get `undefined` as the notification.
This fixes the mismatch to supply both the notification and the action to the registered callback.